# backend-server-examples/php/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "DCID Server SDK - PHP/Laravel",
        "description": "API endpoints for the DCID Server SDK Laravel example. Provides authentication, identity management, credential issuance/verification, data storage, and analytics.",
        "version": "1.0.0"
    },
    "paths": {
        "/api/auth/sign-in/initiate": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Initiate OTP sign-in",
                "description": "Sends a one-time password to the provided email or phone number.",
                "operationId": "34b38b5d94f6d73007a00126ff58bdcb",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "example": "user@example.com"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "example": "+1234567890"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OTP sent successfully"
                    }
                }
            }
        },
        "/api/auth/sign-in/confirm": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Confirm OTP sign-in",
                "description": "Validates the OTP and returns authentication tokens.",
                "operationId": "09b071fb99893147ec9432908145074d",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "otp"
                                ],
                                "properties": {
                                    "otp": {
                                        "type": "string",
                                        "example": "123456"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "user@example.com"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "example": "+1234567890"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Authentication successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "access_token": {
                                            "type": "string"
                                        },
                                        "refresh_token": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/auth/admin-login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Admin login",
                "description": "Authenticates an admin user with email or phone.",
                "operationId": "d2c52accb3e50180cb09a88e686cb9c9",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "example": "admin@example.com"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "example": "+1234567890"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Admin authentication successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "accessToken": {
                                            "type": "string"
                                        },
                                        "refreshToken": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/identity/generate-encrypted-key": {
            "post": {
                "tags": [
                    "Identity - Encryption"
                ],
                "summary": "Generate an encrypted key",
                "description": "Generates an encryption key for the specified DID.",
                "operationId": "5a980d32b7241c7bae271efffa118a5d",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "did",
                                    "ownerEmail"
                                ],
                                "properties": {
                                    "did": {
                                        "type": "string",
                                        "example": "did:example:123"
                                    },
                                    "ownerEmail": {
                                        "type": "string",
                                        "example": "user@example.com"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Encryption key generated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/identity/get-encrypted-key": {
            "post": {
                "tags": [
                    "Identity - Encryption"
                ],
                "summary": "Get an encrypted key",
                "description": "Retrieves the encryption key for the specified DID.",
                "operationId": "95c44494df7e61037a5668f548aa4ff2",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "did"
                                ],
                                "properties": {
                                    "did": {
                                        "type": "string",
                                        "example": "did:example:123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Encryption key retrieved"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/identity/issuer/issue-credential": {
            "post": {
                "tags": [
                    "Identity - Issuer"
                ],
                "summary": "Issue a credential",
                "description": "Issues a verifiable credential for the specified DID.",
                "operationId": "83c479f0019b54773c210a740e6b209d",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "did",
                                    "credentialName",
                                    "values",
                                    "ownerEmail"
                                ],
                                "properties": {
                                    "did": {
                                        "type": "string",
                                        "example": "did:example:123"
                                    },
                                    "credentialName": {
                                        "type": "string",
                                        "example": "KYCAgeCredential"
                                    },
                                    "values": {
                                        "type": "object",
                                        "example": "{\"birthday\": 19900101}"
                                    },
                                    "ownerEmail": {
                                        "type": "string",
                                        "example": "user@example.com"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Credential issued successfully"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/identity/issuer/get-credential-offer": {
            "get": {
                "tags": [
                    "Identity - Issuer"
                ],
                "summary": "Get a credential offer",
                "description": "Retrieves a credential offer by claim ID and transaction ID.",
                "operationId": "e44ded52c6a839b95478cad4a609d515",
                "parameters": [
                    {
                        "name": "claimId",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "txId",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Credential offer retrieved"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/identity/ipfs/store-credential": {
            "post": {
                "tags": [
                    "Identity - IPFS"
                ],
                "summary": "Store a credential",
                "description": "Stores a credential for the specified DID.",
                "operationId": "1df7fdeb4dc83d1f2b81c017f693e41b",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "did",
                                    "credentialType",
                                    "credential"
                                ],
                                "properties": {
                                    "did": {
                                        "type": "string",
                                        "example": "did:example:123"
                                    },
                                    "credentialType": {
                                        "type": "string",
                                        "example": "KYCAgeCredential"
                                    },
                                    "credential": {
                                        "type": "object"
                                    },
                                    "encrypted": {
                                        "type": "boolean",
                                        "example": false
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Credential stored"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Enter the JWT token obtained from sign-in or admin-login",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}